pyproject.toml
[project]
name = "maf-py"
version = "0.1.0"
description = "Microsoft Agent Framework Tutorials"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework>=1.0.0b251120",
    "python-dotenv>=1.2.1",
]
